Relocate PersistenceState type
