+++ src/ADempiere/modules/persistence/store/Persistence/state.ts
@@ -1,11 +1,4 @@
-import { IValueData } from '@/ADempiere/modules/core'
-import { KeyValueData } from '../../PersistenceType'
-
-export interface PersistenceState {
-    persistence: {
-        [columnName: string]: Map<String, KeyValueData<IValueData>>
-    }
-}
+import { PersistenceState } from '@/ADempiere/modules/persistence'
 
 export const state: PersistenceState = {
   persistence: {}
